+++ CHANGELOG.md
@@ -4,6 +4,7 @@ all notable changes to this project will be documented in this file.
 
 # [0.1.0] - 2025-06-20
 
+- (Feat) - CSelectMulti
 - (Chore) - update react-day-picker to latest
 - (Refactor) - update daypicker for more flexible
 - (Fix) - update validation onChange year and mond based min and max date